Move to FileBasedDataSourceSuite
